requirements.txt
pillow==12.1.0
stockfish==4.0.5
